347681: interactive console (has to be demo-able at SW)
first cut of the interactive console that works directly on started clients (without a debug hit) but then in its own view.
You are in the globals scope where you run your commands in. But the creation of stuff like "var x = 10;" will be done in its own scope that builds on globals, and will be remembered also by the code completion.