Fixed case: SVN checkout causes loads of invalid warnings.
There were 3 main problems:
    1. receiving changed events for partially written - thus invalid solution projects
    2. building and caching this invalid stuff for build while checkout was still going on
    3. not receiving last batch of changes (last 2-5 projects) that should have come from SVN checkout.
    
(1) resulted in exceptions here and there in places where null/invalid solutions were not taken into account. (fixed)
(2) many warnings remaining behind because of outdated build cache; partially because of (3) (build cache was also generated for non-active solutions) (fixed)
(3) servoy project list not refreshed correctly or if it was, the solution metadata for the last projects was not added to the repo. so they were unuseable => more problem markers. (fixed by not deserializing solutions when not needed in change handler, so much less activity)
Case: SVY-1013