md5 to safer hash (first cut)
added PBKDF2.jar to the classpath of shared and added it to the jnlp servlet

created methods on or Utils class that creates a hash and can verify it. 
Made javascript methods in the JSUtils class that calls those so that developers can use those to generate hashes.
Case: SVY-319