Implemented: SVY-1711 Have an option to run at least unit test for a specific form instead of the whole solution. and SVY-1333 Add option to run a single unit test method.
You are now able to run JS unit tests the either on whole active solution, or on a form, on a global scope, on a module of the active solution, or directly on a method.
Also solved a NullPointer when starting the test client before starting a debug smart client + a few changed related to service provider.
Case: SVY-1711, SVY-1333