Deadlock when importing solution in developer
This job should be used when a workspace job should be run without any concurrent resource change listeners fires when this job runs.
So instead of using a {@link WorkspaceJob} instance a {@link IWorkspaceRunnable} should be created and this job should be used to run it.
The same configuration should be used with the {@link ISchedulingRule} that should just be set on the Job and it will be transfered over to the call {@link IWorkspace#run(IWorkspaceRunnable, ISchedulingRule, int, IProgressMonitor)} with the flag being a {@link IWorkspace#AVOID_UPDATE} to block the eclipse NotificationManager to run its NotifyJob concurrently.

Case: SVY-3122