For ongoing SVY-3291 : JSDoc was not presented properly for JSList and JSInsetList.
Before, for a Method of a class (not interface), the annotations were searched for starting with the declaring class method. This meant that when class A extends B re-declares/overrides the method in parent class, the annotation searching was performed starting from A, in all interfaces and extended classes. But if you had a method implemented only in B, it would only search for annotations starting in class B and it's interfaces, even if you were interested in class A's JS API. Because of this, if A implemented an interface which annotated the method in B to be seen in JS, it would not be found (A's interfaces were not searched in).

Now the searching for annotations in classes/interfaces always starts from the original class (A).
Case: SVY-3291