performance issue in servoy 6.1.4
cache everything on Type (type itself but also the Property or Method members on their types)
Property must be on its own cache because the properties value also gets a declared type (dltk changes the value object that we return, have in cache)

now if you do 2 times in your js file:

globals.

then that are 2 properties but they now map to the same cached ValueCollection so they don't have to be over and over again build up.

Case: SVY-4063