3 bugs in my solution (7.0.1)
for point 2 it was very hard to see which element really had a problem
because those elements don't have names then we just say that there is a element with a problem on that form, but you have no idea which one it really is. So just fallback to uuid so that users can see or search for that. 
Case: SVY-4228