Date Picker needed
support for: http://amsul.ca/pickadate.js/v3/
there is a small patch in the picker.js file for this bug: https://github.com/amsul/pickadate.js/issues/120

DataCalendarField has support for toggling between native pickers and the pickadate from above.
Currently if in touch mode then we show native pickers. But this should be enhanced for specific browser versions i guess.
If using native pickers then date formats are not supported at all, how it is displayed is browser specific, and over the line it is always in "yyyy-MM-dd"

When the bug is fixed in the js lib we should use the minimal versions.

Case: SVY-4400