For SVY-4441 : on exporting solution, dbi files are ignored
If DB is down, ignore error markers just as before and export.
If DB is online, check for error markers in resources project as well. Only ignore the resources error markers if "use dbi" is checked. Otherwise error markers in any of the modules or active solution (just as before) + (now) the resources project will not allow export.

Improved the export solution wizard messages a bit as well.
Case: SVY-4441