Fixed a regression introduced by SVY-3941:
Avoid the creation of 2 ServoyModel instances by the same thread (reentrant).
Also if this is detected in the future - throw an exception and handle it correctly both in case of servoy model async and sync init. Still, allow servoy model attempts to initialize in the future. This will make the problem more obvious then just noticing strange behavior with nothing in the log files.
Case: SVY-3941