Render actual icon for media entries in the Solution Explorer when the entry is an image +add chache for SWT images grouped by folder
cache only 1 folder (because its a list thing and for that list thing it is constantly asked for)
reusing/disposing done a bit different.
only call dispose/create a new map when the IMAGE type is asked for, not if it is just folders.
SOLEX.dispose() should also dispose the images.
Case: SVY-541