redirect the solution.js file to a servlet/service running in developer when exported in debug mode
all the extra properties are now set as query params on the .js and the json.js file
Also added the solutionName itself to it so that they map really on the right solution even if for example the service solution is live.
also the special webserver will also go up if a server solution is loaded that has a shared mobile module.
Case: SVY-4989