Implemented SVY-5460 : Unit tests do not fail if onLoad handle method in a form throws Error.
When there is no onError handler registered at solution level or when the error handler returns "true" (which means "I can't handle this error, report to user), errors that are reported to that handler will now fail the current test (after the test method's execution finishes - so it doesn't stop method execution as normal clients don't do that either). If the handler exists and it returns false or doesn't return anything, the test won't fail because of that error as the error is considered already dealt with.

Also refactored a bit to prevent some previously existing hidden dependencies, prevent code duplication and localize more of the test code in one project.
Case: SVY-5460