have code compete work for webcomponents (based on api from spec)
properties: removed the 3 that we add like size,anchors and location, also dataprovider properties are not added (because what are they then, the string of the dataprovider or the value..)

api: added the methods what we currently support, so with optional support but no doc. As types for return and params only basic type support (string,boolean,number,date). No mapping yet for the  complex types.
Case: SVY-6306