SVY-6877 (multi) selection
first cut, single and multi select (with ctrl) works, also deselect from
a multi select. 
It directly selects on mouse down when the node didn't have a selection,
if the node already was in the selection it only selects that one after
mouse up if the mouse was still in the same position.