For SVY-7864: automated export modules list is not working
More descriptive hint on what to check.
Do not check error markers for import hooks.

Removed code that introduced the problem initially because it expected
modules list to contain import hooks as well. That code made it still
execute the export and then fail after successful export. It should
either fail fast or let the real exporter code fail it (like it will be
now).