For SVY-6474 : Unit tests with syntax errors do not fail when they are run through Jenkins
More desctiptive hint on what to check.
Do not check error markers for import hooks.

Removed code that introduced the problem initially because it expected
modules list to contain import hooks as well. That code made it still
execute the export and then fail after successful export. It should
either fail fast or let the real exporter code fail it (like it will be
now).