Fixed SVY-8934 : can not set foundset to null in component of derived form.
Moved the null conversions to WebComponentPropertyHandler - and they
only happen when needed so as not to disable null usage in object/array
types.

Fixed a problem where nested object values that are JSONObject.NULL
would not be written at all by toString - because it was using some to
java wrapper maps there.

Fixed another problem where WebCustomTypes and their underlying json
would get out of sync when setting a subproperty to null - the result
being when a sibling changes in the future the previous set to null is
lost.