Implemented SVY-8589: Dropping child components from palette into custom web components
Previously there was no support for child web components/component
property type in developer. Fixed various problems discovered in the
process.

Also fixed a problem related to case SVY-8934 (it was reopen because of
it) - restoring to default a DP subproperty of a custom object property
gave a class cast exception. (was trying to set JSONObject.NULL instead
of null)