SVY-10140 Script editor became very slow after a while.
seems to becaused by the refreshes of the browser/editor parts when they
are not visible. That is a performance problem anyway, because we only
really need to refresh when the editor is made visible again. So now
just set a boolean and delay the refresh until this editor is made
visible again.