For SVY-9493 : Ability to configure what ends up in the Developer console
Always enable in 8.2 server log forwarding to debug client console
(console view in developer) when starting developer; automatically puts
the appender props. and adds it to root category before initializing log
system. And don't leave it in root category in properties file as it's
not to be used when running app. server.

For now I leave the appender properties themselfes in the properties
file just in case someone wants to tweak format for example.

Still if someone gets annoyed by these new messages (which I am not sure
anyone should want) they can be disabled by this property in the
servoy.properties file:
"servoy.developer.disableServerLogForwardingToDebugClientConsole".