Related to SVY-11409
Fixed another problem with form inheritance and custom object types.
If you would set a value in parent form and then change the value in
child form of a custom object root prop. of the component, then reset
would not work correctly on that property.

Also resetting that property once it started to work resulted in a class
cast - because the inherited-from-parent custom object persist was not
recreated => it gave a JSONObject instead of WebCustomType for that
property.

Also setting an int array to it's default values manually (so without
reset) would not identify correctly the situation.