SVY-12067 Make the websocket part of nglclient2 work (so an NGClient starts up)
make a special build_debug script that makes a buidl with source maps
and watches the dir. That script is run for ever then in the developer
auto picking up changes.
(handy for our own debugging, now we also copy all the time changes we
make to the ngclient.ui/node dir and later on when we really modify the
output dir because of service or component changes)