Fixed as part of SVY-13259.
The actual bug described in that case is no longer reproducible due to
fix from SVY - 13258. But when testing I found there was another
problem. If you have two forms and in next commits pushed by someone
else maybe form1 is deleted and form2 is renamed into form1 and you pull
it would end up with duplicate form1 persists in the solution.

Then there was another thing that needed to work - if when you pull
form1 would be renamed into form2 and form2 into form1 then it ended up
only showing the new form2 (form1 was deleted as a persist from the in
mem. solution although it is there still on disk).