Fixed SVY-13962: War export wizard next button does nothing
Actually next button should have been disabled there. We were not
following the contract of canFlipToNextPage and using it as a substitute
for page complete - even when computing finish.

Other fixes/enhancements made along the way in exporter:
- command line equivalent copy to clipboard enhancements
- allow data model changes UI enhancements
- doNotOverwriteDBServerProperties instead of
overwriteDBServerProperties in command line export - in order to have in
both UI and command line the same default; this is a BREAKING CHANGE
that should be mentioned in release notes