SVY-14423 Test big import (and building),why it is so slow (creating many times the PersistIndex)
if we are looping over outstanding changes then in every loop the
modules of the active solution shouldn't be flushed and recreated (the
editing solution) this should be done once when everything is finished.

I don't get why this outstanding changes and file changes are hitting a
lot in this area. Because to me an import should just be a write on disk
then an import of the project (just like a start of the developer) not
that then all kind of changes are being recorded..