SVY-14448 split up the ui.tweaks plugin into 2 plugins
introduce a special plugin for just the weaver code
the actual code that resolves the url is still in the old plugin that
has an activator
That plugin is trimmed down to have less depedencies (because it is
loaded very quickly) So it is not a UI Plugin anymore it doesn't depend
on eclipse.model (so no is running in ui check...) and the Pair classe
is copied from shared.

The new weaver plugin is tied to the old tweak plugin only by the
wovenClass.getDynamicImports().add("com.servoy.eclipse.ui.tweaks");

that makes sure that the ImageReplacementMapper can be found there.