SVY-14423 Test big import (and building),why it is so slow (creating many times the PersistIndex)
made an extra indirection to the ServoyModel so that that
DelegationModel can block any creation or loading of solution so it is
fully controlled by us. (doing it in the background)
all callers of getActiveProject() or getServoyProjectsAndModules()
should be able to handle a null or empty array.

The editors are now lazy based on the above, they will only fully
initialize themselfs after the active solution is set.