For SVY-14617: Add support for tooltips on a webcomponent's properties in designer
Changed the impl. so that it takes the tooltip directly from the
propertydescriptors of the tree items (this way we don't have to rely on
current selected component in editor, so in case of nested structures it
is sure to go to the right level).

Added support for html tooltips similar to what solution explorer has.

Added support for design-time-docs for legacy components (taken from the
xml that is also used to generate wiki page). IMPORTANT: for now that
xml is manually built (mvn install -Pdesign_doc on the servoy_client git
repo - Jenkins does not do that automatically) and committed (to
com.servoy.eclipse.core.doc/servoydoc_design.xml)

Both handlers and properties in spec now use "doc". Before one used
"description" key and one "tooltip" tag.

Array elements are also supported now in regards to "doc" tag.
Layout container props. supported.

Dltk tooltips in script editor also show this documentation for custom
component properties. Same with SolEx tooltips for elements under forms.

Also fixed two exceptions that got logged if developer is offline (we
shouldn't spam exceptions then, it is a valid situation).