For SVY-14617: Add support for tooltips on a webcomponent's properties in designer
Added "doc" to schema and made elementConfig better (also used $ref to
not duplicate element config definition in model and types).

For the future - when the json .spec editor we use supports JSON Schema
version 2019.09 (also known as 8) - we can remove duplicate code even
further by using "unevaluatedProperties": false instead of
"additionalProperties": false (in combination with "allOf") in order to
merge the propertyConfig with "type" and "elementConfig" keys in order
to create the object type definition of properties and types. Currently
there is a java JSON Schema validator that supports 2019.09
(https://github.com/networknt/json-schema-validator) but not sure when
that will be available as an eclipse plugin and in which plugin.