SVY-14768 Solve errors on servoy_sample_ng_basic
added an extra boolean so that we don't do a resetUUID when we are just
loading in the solution. This can't happen when just loading in the
solution because then we only adjust the thing in memory that is then
different then it is on disk.
We should only reset the uuid when we know we will then generate the
file on disk