Fixed SVY-14881: Missing ng services from SolEx. after startup
It was a threading issue - and possible (but unlikely) deadlock
situation. Plugin nodes for solex were loaded in a background job while
the active solution got activated. As the active solution contained
ngservices that go in that node it was resetting all children in the
middle of the old children getting created (background load job was
trying to load some images in the SWT thread, but SWT thread was busy
activating the solution; so background loop on old services list was
waiting for activation which reset the plugins node but then background
loop would populate child nodes based on old services).

Now we only load the java plugins in background - which don't need SWT
thread - and are not affected by reload due to ng services changing
anyway...

Also fixed a ConcurrentModificationException (listeners got added on one
thread while being fired on another thread - BaseNGPackageManager).