SVY-15207 Creating calculations by hand result in null pointer or class cast exceptions
fix the null pointer when calculations are just created (and don't have
a name yet). We should double check if the datasource cache is created
correctly after that.
Also when the null name was checked then a class cast exception because
the code always assumed the persist was a tablenode, but it mostly isn't
(its mostly a ScriptCalc or Aggregation)