SVY-15519 [NG2] port or check servoyformat.ts and look at the removal of numeric
added formatting.service that does te formatting (not the parsing)
that is used by the FormatPipe and by the new format control value
accessor directive that replaces te format.directive
FCVA can be used be field with NgModel and the svyformat attribute
this class writes the model value to the view (value property of the dom
element) by formatting the model value and parse this value back from
the dom element into the model.
Can be later used also for validation.