SVY-15556 [NG2] make a test solution that test all components of the default and bootstrap handlers SVY-15546 [ng2] listformcomp - check what remained to be implemented
generate for a Cell a special Model object that extends the model object
of the ComponentModel.model for every cell in a row.
Cache this Cell in the ViewPortRow cache.
implement getHandler/datachange/getServoyApi callbacks from the
compopnent templates
Reroute the startEdit and apply to the ComponentModel.state.
"config" form components attribute is renamed to "svyContainerStyle"
removed the not needed foundsetChange emitter
FormComponentType -> FormComponentState
removed "window" from "window:keydown" only keydowns in the component
should be working not of the whole window.
implemented event.viewportRowsUpdates foundset event (to push the
changed modelViewPort values to the Cell model)
changed how selection class works for the row div
use the new component templates that now needs a callback context
Don't return the viewport rows yet when the number of cells are not
calculated yet.
calculateCells should use the test for isAbsoluteLayout() of the parent
form where it is in and report an error when the responsivePageSize is
not set for a responsive parent form.