SVY-15546 [ng2] listformcomp - check what remained to be implemented
register of the actual components that are on a form (component) so that
the form component can call api on them when it is called for.
For this we need to call servoyApi.registerComponent(comp) in
basecomponent this method is empty by default and FormComponent and
ListFormComponent are implementing those to store that component by
name.