SVY-15556 [NG2] make a test solution that test all components of the default and bootstrap handlers
when a model property is pushed and this is a "smart" proprety (has
conversions) and when the value really doesn't change compared to the
model reference itself. do make sure the component knows its property
did change.
Angular by default doesn't do nested (deep watch) checks, so this is a
replacement of that.