SVY-15738 introduce "clientfunction" property type to avoid doing "eval(model.xxx)"
introduced "clientfuntion_converter" for the property type itself (gives
back a functio)
introduced "clientfunction_service" that handles the loading of the
server side generated clientfunctions.js and reloads
Als this service has a function that returns a promise when the system
must wait for it. which window.service.updateController nd
servoyServices.loaded() uses to block the form show for a while.

slider can now just assign directly the value as functions no need to do
"eval()"