SVY-15738 introduce "clientfunction" property type to avoid doing "eval(model.xxx)"
wait for the loaded signal of servoyService (that also test for
clientfunctions)

doing a reload when already showing a form is very early then the main
form doesn't need to be there yet. so test for that and also test if the
component that is used to show a form is already there before trying to
show it.

code clean ups.