Dataproviders should do "apply/svyPush", other property "dataPush"
the template will generate for dataprovider type callbacks an extra
boolean that says this property is a dataprovider and the emit() should
do svyPush instead of dataPush.

This is also give to the ListFormComponent and then relayed to the
Component type. That is now sending only the properties that are marked
as changed