SVY-15913 build and test Servoy 2021.03 rc1
make the karma test run in the pom.xml a variable so you can override it
add the extra nowatch Edge test run in the package.json that then can be
used.
This way people can decide when making an installer what browser should
be used to run the test.