SVY-15741 [NG2] create a first E2E test that runs and test NG2
added Testability service to the websocket messaging so that e2e scripts
waitForAngular() do wait until the messages are resolved from the
server.
added a createWSEvent(): Deffered method to websocket/sabloservice that
components should just to create a deferred id and a promise that makes
the e2e script wait.