SVY-16029 [NG2] investigate how services can be there own project (and included in ng2)
big refactor to create a servoy-public lib that can be used by services
or component packages.

fix a lot of test to include now a more standard servoytesting.module
that has the basics overwritten and injected.
Fixed a lot of test to include the right modules so all directives are
found.