SVY-16029 [NG2] investigate how services can be there own project (and included in ng2)
install the services if needed
make sure npm install (wth services or without) is always done.
npm build will not always be triggered only on change of the  source or
if the dist folder is not there at all.

the binary build of nclient.ui will not have a dist folder anymore