SVY-16367 Ant (servoy-cloud) war export isn't working for NG2.
move copyNodeFolder a action that needs to be called on the ui.Activator
instead of the activator already do that on activation/start
Because we should be able to override the Console where the log should
go to, this is now behind an interface.
checking if a default package should be ignored is now really relayed to
the active package manager (and the exporter always says never ignore)
WebPackageListener can now be set in production mode. This does mean npm
install and npm run build will be always done.