SVY-16536 Update Java and NodeJS
update the shipped Java to the 1.7 release of Hotspot
They are now created with the create_jres.bat files of the build
project.
And the generated jre's are upload to S3, all are now just zips and
those are packed in the root, so the download/unpack in these pom files
are much easier now. (now special always changing dir name inside the
tar or zip)