For SVY-16500: [ng2] viewport implementation does not take smart values into account for notification
There was some obsolete code in foundsetlinked that had to be removed
now that viewport handles smart values correctly, but when testing that
with a list form component and custom valuelist/global variable on a
typeahead it still didn't work - because a similar piece of code was
missing from component type - to handle smart values in non-viewport
properties. So I added that as well.

Now I suspect that component type (child props) still will not work
correctly for non-change-aware/dumb properties. So if component emits on
such a prop I don't think it will reach the server.