SVY-16599 [NG2] ngutils.addFormStyleClass() can handle $ in the name (uses a selector)
don't reuse the form classes of before (by concatting the current
styleclasses on the previous value)
this is because the name change of a form should be seen as a constuctor
call. everything should be reinitialized because it is a new form (the
form that is shown goes from A to B). No state should be kept.