SVY-16660 [NG2] - application does not reload after a new build has been deployed
onconnecting was out of sync with the NG1 implementation, its now made
in sync with it. So we only look a the reason to do really stuff (like
reload or calling websocket.close())

also make sure that the event emitter that fies if it is in reconnecting
or not is done in the ngzone zo it it seen in the ui.