SVY-16968 Changes not being saved/ showing in GitKraken
call createIndex() now when the developer index is created
won't happen anymore in the super constructor.
fix the duplicate problems, we can't do equals because that is just
based on uuid which is the actual problem
same for adding in to the list of duplciate we can't do contains because
that would be true. But we should be able to always add both at init of
that list and just keep adding the new later..