Fix for SVY-17091: Rename of media file is creating a copy
Editing solution PersistIndex was not updated properly.

PersistIndex listens to the solution for changes
that means
    EditingSolution persist index listens to changes on editing solution
    Solution persist index listens to changes on repo./proj solution

RenameMediaAction did:
- changes the editing Media persist (isChanged == true in the perist)
- calls project.saveEditingSolutionNodes(Media)
- ends up in EclipseRepository.updateNode that in the end does
solution.clearEditingState(processed)
- isChanged becomes false on the Media persist from editing solution
- NO CHANGE HANDLER IS FIRED ON THE EDITING SOLUTION, although Media
persist was changed and saved to disk
		<<<<<<PROBLEM IS HERE>>>>>
- PersistIndex is not aware of any changes because the listener did no
fire
- then ServoyModel resource change listener takes over, it updates the
solution (not editing one) and fires it's change handler so the persist
index of the non-editing is fine
- then it updates the EditingSolution persist based on the changes from
disk - but that one is already changed (names match, isChanged remains
false on the editing sol. Media persist)! so no change handler fires
there either. Method
- ServoyModel does ServoyProject.updateEditingPersist ->
copyNodeToEditingSolution that would do
dest = ((AbstractRepository)editingSolution.getRepository()).copyPersistIntoSolution(src,
currentParent, hasChilds);
that would normally trigger the persist change handler to update the
PeristIndex of the editingSolution - but if it has changes; in our case
it does not have changes