SVY-17291 support adding a responsive container in a responsive container
changed layout containers back to the map (to what it was)
and don't use that anymore in the designer
but make sure that the parts components can now also be parts structures
(layouts) that are the root of the part. So now they are rendered in the
same loop as the normal components.