Related to SVY-17008: [Ti] Scrolling LFC & datagrid issues
Make sure that the handled req. property does not end up in the foundset
value that clients use - in case we received a full value from server.
Also make sure that if the handled requests is the only thing we
received from server, it is not interpreted as a full value and it
doesn't blank out the foundset prop. or give errors because of that.