SVY-17284 Building process every time when switching solution - Servoy 2022.6 (Final) - [NG2]
changed the way the target folder is
now every solution has a target\solution target folder for the node
stuff
the \target root folder is just for the main package.json that is not
installed into anymore.
the node folder has now a timestamp scanner to see if something did
changes in our core resoces
also the dir packages (source packages) of components now do that same
scan to check if only some source is changed.
if one of those are changed then a full install/compile will happen.