SVY-17727 variants rewrite to use a json file that just is a set of classes
removed Variants stuff completely from the ServoyModel and its
interfaces
because now Variants is a runtime (servoy_shared/flattenedSolution)
thing. Because it needs to resolve also at runtime.

now instead of pushing a styleClass property to the model of an
component (based on variant) that needs to be created
we need to push the variant name on the variant property

Except when showing the variant popup, then the variant property must be
directly the classes of that variant (and not the name)
this is because that is a runtime representation

So Runtime representation, the variant property of a componetn is a
array of clasnames
at designtime, persist property, the value is the name of the variant

the VariantPropertyType will convert that.