SVY-17678 Copy & Clean build for TiNG in developer doesn't start (visually) and can run multiple times simultaneaously
Added output at the beginning of the full delete (if you do a clean) and
report the total time for that.

Made all the build jobs (Delete, CopySources and the Actual Build job)
all belong to a specific family.
If there are jobs running under that family then the
CopySourceFolderAction is disabled.