SVY-17678 Copy & Clean build for TiNG in developer doesn't start (visually) and can run multiple times simultaneaously
when installing a solution from spm, we also should not run it multiply
times at the same time.
Also not for the default dir that was "solution"
now the default dir is null and the callers can handle that.
also a check in checkPackages if there are already jobs running for this
family