SVY-17621 Issue with dataprovider type lookup for format properties in Webcomponent spec
look up the right format type.
If it is a nested WebCustomType then first ask it on that webcustomtype
(if it has a dataprovdier property for example)
and then go to the parent of that custom type (which can also be a
custom type).
Until we hit the WebComponent itself.