SVY-17937 custom type of a zip package overwrites the src package one
if we hit 2 or more readers with the same package name
and there is already a "Container" reader then don't add this zip one in
the packages. Container readers should just always overwrite them all.

The problem is a bit what todo if there are 2 different zip readers of
the same package. Because what is used to compile the sources and what
is used to export the spec files could be different
(but there should be a warning if the version doesn't match up)