Another exception not necessarily related to SVY-17930
Gabi noticed that after a browser refresh with the same sample in the
case, but using not a filter but and edit type-ahead, an exception
happened because the foundset linked non-record dependent valuelist was
an array of size 0. That happened because the updates from server
contained both the columns and the foundset property and the columns
were handled first (while foundset was still undefined) => it didn't
generate the initiall array of viewport size size.