A fix for SVY-17946: ListFormComponent sorting and duplicate data issues
A child component linked to foundset would generate in model_vp
ModelInSpecificRow instances with the prototype set to main model (for
non fs. linked properties of that component). And a proxy on the
ModelInSpecificRow would decide if it needs to set a property in the row
object or in the main model (prototype). But it seems that Reflect.set
on the row ends up calling the proxy in the prototype (I expected it to
just set the value...). So we want to avoid the proxy of the main model
(non-fs-linked properties) to assign foundset linked properties to
itself due to it's own proxy (if the component has Proxy in it's model).

+ a big code comment out (proxies for models of components were not
really useful as the root props can be changed by ref in the model only
if an "emit" is called on the apropriate property @Output, and we do
what is needed in that case (send to server/add listeners)