For SVY-18164: Changes are not sent to server for a custom object after that custom object is sent as arg. to a handler as well
Added unit tests and found another (but similar) situation that needed
to be fixed: if a custom object/array from the model is changed (push to
server allow for example), then, before being emitted it is given as an
argument to a handler call, then it's changed state would be cleared by
that and the model change would not longer be sent to the server (even
on emit). So basically if it's an arg or return value it should always
send full value, but never clear the change flag (changed key/elements
or fully change flag).