SVY-18025 Console.log statement locations in code are not easy to find due to override from sablo_app.ts
refactored it a bit to use proxies and have only 1 function that does
the callservices

but the bigger change is that all levels of console except error are
mapped to "warn" because warn (and error) do generate a full stacktrace
that you can expand and that is also a source mapped stacktrace.

Error maps just to error and also error will give the stack trace (of
new Error().stack) to the server so there errors will have the stack
also. (In production this is harded to read because of the minified
stuff)

when you are in the browser and you want to go back to default logging
there is now a function on the window object that does that:
toggleSabloLogWrapping()