For SVY-18378: Spike: Research & developer Servoy-tuned LLM
Added partial code for upserting similar to what worked with gradle. But
it currently does not work as there are various build system issues; we
have to use maven (so what gradle generates) wrapped inside generated
osgi bundles from the target platform (the openai libs export the same
package in multiple jars, there are some deps. to android from what I
can tell that are "provided" - not sure what to do with those, it
doesn't find a DNS resolver etc).