SVY-18344 Input Control/Data Label - immediately choose data provider
refactored so that the properties view and the create/add components,
main webcomponents or custom types are all done through the same code.

What table to use is calculated through the same code for properties or
the "wizard" dataprovider dialog
CustomTypes and Webcomponent uses the same show dialog code to look if
there are more then 1 property, if there is more, then first a dialog is
shown to ask which properties to set
This dialog auto selects "dataprovider" names because that is the most
used one as the main.. (bit hardcoded)
Als "ID" is filtered out of the name because that is also done for the
properties view.