SVY-18418 Issues with properties in developer
a property value of a webcomponent can be null for 2 things
1> its actually a null value in the json
2> the json map has no key for it (kind of undefined)

but this is a difference because when something is <2> then the default
value of the property kicks in, but if it really null then its really a
null value.

But for now the value that the properties view gets are just null so
labelproviders needs to somehow check themselfs.
Also when setting we really need to just set a null value because it can
be that this is really a undefined to null thing.