For SVY-18502: LLM: collect training and test data
Finished the web comp/service part of generating documentation for
pinecone embeddings (we could use the same to generate fine tuning
data). But as pincone similarities are not enough for finding specs
needed in code snippets (>1800 entries in the index gave completely
useless 10 top entries), I am thinking that maybe fine-tuning doesn't
need this preprocessing necessarily; we should I think first try to
teach it what the .spec file is and see if feeding it raw spec and jsdoc
files will be enough or better.