SVY-18890 upgrade to angular 17.x and also change the builder to the "application" (esbuild) builder
the public api (servoy/public) is not found anymore when a package is
installed as a package (zip) instead of directly from source.
see: https://github.com/angular/angular-cli/issues/27176

now we need to build and then install the public package.

This has a drawback that we can't easily debug through thet package
anymore and develop it (it must be source for that)

will need to see how we can make this part of the full source again and
if we can at least make a distribution on jenkins so it is included
directly in the build (we don't need to build it all the time)