SVY-19334 updater doesn't run when updating a version to 2024.6.0_RC1 (and web.xml is not updated or in the update file at all)
test and create a "version.txt" file when checking for an update.
If j2db.jar is not found anymore to get a version
then check for a version.txt
If that one is also not there then assume this was a clean install and
create quickly a file with the current version in it.

So the next time (when an update is done) this file has the old version
number.