Fixes for SVY-19635: Foundsetpush issue
I noticed a few problems when debugging, many of which would fix the bug
by themselves - if the were corrected.

Problems and fixes:
- on initial form show, toJSON would be called twice on props of a form
when calling updateController, only one of them being used; fix: it is
called and used only once now.
- on initial form show, after calling updateController (that does full
toJSON) it would still think it has changes on that form  and send again
a bunch of n: true or other non-changed values to client in the same
message. Fix: it now clears changes and does not need to send the no-ops
in the same msg. to client.
- updateController would send to client values before the form would
even have data initialized; so not even with data not just before onShow
handler. Fix: now updateController content is generated after data is
present in the form and after the onShow solution handler is called.
- the values that onShow solution handler would modify would be sent to
client in a separate message, after updateController, and possibly  be
applied between updateController and the updates that were in the same
message as updateController, resulting in a wrong final state (because
some wait for the form angular component to show up). Fix 1: instead of
having these 4 toJSON of which 3 reach the client and are applied in
various orders, now there is only 1 updateController to JSON in most
cases containing the state after onShow. Fix2: client side,
updateController happens before data updates are applied now, fixing the
previously broken order of data apply. Fix 3: client side, if we get a
data update for a form, we don't care if that form is visible (has an
angular component), we only care if the formCache (the state/model) was
initialized by updateController; if it is so, then updates get applied
then; if it is not, a warning message is sent to browser console + the
unexpected data update(s) will be processed sequentially after
updateController for that form has initialized the form's formCache.
- when running e2e tests, the needed data-cy attribute would be missing
if the form's onShow would change the attributes (so at runtime) of a
component. Fix: now the data-cy attr remains in place.

I also re-enabled the angular runCoalescing: true optimization that made
this bug surface easier.