SVY-19651 create a first angular based offline (mobile) client
we introduced a special environments file that is a special build for
mobile
2 extra scripts in the package.json to build mobile for production or
debug (using that new environments file)

That environments file now gives a "mobile:true" propery that is used in
a few places todo something different then in a normal TiNG run
Like not creating an actual websocket but the new MobileBridge

Also the window.service.ts is using that booleanm to check for the
special _formdata_ to create all the form caches.

The locale.service.ts uses this to do a bit differnet loading of the
angular locale files.