SVY-19820 Handle bigints (java/javascript) to be able to send to the client (json doesnt directly support this)
added bignumber.js as a npm package
this can handle large numbers and can also do some formatting
When we notice such large numbers when formatting or parsing we will use
the bignumber stuff instead of Numbro.

Numbro seems to have support for this (it does ship bignumber with it)
but it doesn't really fully use it when parsing or formatting.