SVY-20353 build and test Servoy 2024.3.x lts
removed the wrapped http componnts and really use the one Eclipse ships

else problems when installing stuff (like github copilot) will happen.

Also moved the egit/jgit/marketplace and tm4e to the required features
not included features so that versions are not fixed.

updated the products update sites to the correct eclipse once that this
build is build on top of