SVY-20345 investigate/implement if we can use the servoytheme.less.properties in a form.css file
new solution now generates the new one of less
So that there is a pure properties.less for the solution that can be
reused.

Adjused the properties editor so it can handle that file, but also can
handle no theme properties (just there own)
use the ExtensionBasedTextEditor just like the form less editor for
nicer coloring.