Fixed a bug noticed when merging code for SVY-19051
If in a responsive form I have a row with 2 columns; each column has a
button. I then drag button 1 in column 2 but before mouse up I press and
keep pressed the CTRL key (careful not to move the mouse after that),
then I release the mouse click => visually I see the button moved to the
second column but a save and reopen of the form shows the initial state.

It was because it was calling 'create' with an argument constructed for
'move' in the same method.