SVY-20687 Autocomplete on foundset doesn't return the correct fields (new fields)
now the clearCache() is called way more often it was shown that many
jobs where created because runCearCacheJob was created a job that then
calsl clearCache() which itself als creates a job again

removed the runClearCacheJob() and only let clearCache() being called.

But there is no need to call another if the previous one didn't run yet.
Else now with a lot of tables change fires it can trigger many many
jobs.